Factor out the search component so that it can be reused
