+++ apps/src/lib/levelbuilder/lesson-editor/ResourcesEditor.jsx
@@ -1,11 +1,9 @@
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import {resourceShape} from '@cdo/apps/lib/levelbuilder/shapes';
-import _ from 'lodash';
-import Select from 'react-select';
-import 'react-select/dist/react-select.css';
 import color from '@cdo/apps/util/color';
 import AddResourceDialog from './AddResourceDialog';
+import SearchBox from './SearchBox';
 import Button from '@cdo/apps/templates/Button';
 import Dialog from '@cdo/apps/templates/Dialog';
 import {connect} from 'react-redux';
@@ -73,74 +71,15 @@ class ResourcesEditor extends Component {
     };
   }
 
+  onSearchSelect = e => {
+    this.props.addResource(e.resource);
+  };
+
   constructResourceOption = resource => ({
     value: resource.key.toString(),
     label: `${resource.name} - ${resource.url}`,
     resource: resource
   });
-
-  /**
-   * Debounced function that will request resource search results from the server.
-   * Because this function is debounced it is not guaranteed to execute
-   * when it is called - there may be a delay of up to 200ms.
-   * @param {string} q - Search query
-   * @param {function(err, result)} callback - Function called when the server
-   *   returns results or a request error occurs.
-   */
-  debouncedSearch = _.debounce((q, callback) => {
-    const searchLimit = 7;
-    const params = {
-      query: encodeURIComponent(q),
-      limit: searchLimit
-    };
-    if (this.props.courseVersionId) {
-      params['courseVersionId'] = this.props.courseVersionId;
-    }
-    const query_params = Object.keys(params)
-      .map(key => `${key}=${params[key]}`)
-      .join('&');
-    const searchUrl = `/resourcesearch?${query_params}`;
-    // Note, we don't return the fetch promise chain because in a debounced
-    // function we're not guaranteed to return anything, and it's not a great
-    // interface to sometimes return undefined when there's still async work
-    // going on.
-    //
-    // We are including the X-Requested-With header to avoid getting a 403
-    // returned by Rack::Protection::JsonCsrf in some environments
-    fetch(searchUrl, {headers: {'X-Requested-With': 'XMLHttpRequest'}})
-      .then(response => (response.ok ? response.json() : []))
-      .then(json => {
-        const resourceKeysAdded = this.props.resources.map(
-          resource => resource.key
-        );
-        const resources = json
-          .map(resource => this.constructResourceOption(resource))
-          // Filter any that are already added to lesson
-          .filter(resource => resourceKeysAdded.indexOf(resource.value) === -1);
-        return {options: resources};
-      })
-      .then(result => callback(null, result))
-      .catch(err => callback(err, null));
-  }, 200);
-
-  getOptions = q => {
-    // Only search if there are at least 3 characters
-    if (q.length < 3) {
-      return Promise.resolve();
-    }
-
-    // Wrap the debounced call in a Promise so we _always_ return a promise
-    // from this function, which resolves whenever results come back.
-    return new Promise((resolve, reject) => {
-      this.debouncedSearch(q, (err, result) => {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(result);
-        }
-      });
-    });
-  };
 
   addResource = resource => {
     this.props.addResource(resource);
@@ -174,6 +113,17 @@ class ResourcesEditor extends Component {
 
   handleNewResourceDialogClose = () => {
     this.setState({newResourceDialogOpen: false, editingResource: null});
+  };
+
+  constructSearchOptions = json => {
+    const resourceKeysAdded = this.props.resources.map(
+      resource => resource.key
+    );
+    const resources = json
+      .map(resource => this.constructResourceOption(resource))
+      // Filter any that are already added to lesson
+      .filter(resource => resourceKeysAdded.indexOf(resource.value) === -1);
+    return {options: resources};
   };
 
   render() {
@@ -215,14 +165,11 @@ class ResourcesEditor extends Component {
         <div style={styles.resourceBox}>
           <div style={styles.resourceSearch}>
             <label>Select a resource to add</label>
-            <Select.Async
-              id="resource_search"
-              name="resource_search"
-              loadOptions={this.getOptions}
-              value={this.state.searchValue}
-              onChange={e => this.addResource(e.resource)}
-              onValueClick={this.addResource}
-              placeholder={''}
+            <SearchBox
+              onSearchSelect={this.onSearchSelect}
+              courseVersionId={this.props.courseVersionId}
+              searchUrl={'resourcesearch'}
+              constructOptions={this.constructSearchOptions}
             />
           </div>
           <table style={{width: '100%'}}>
